Entities can now be referred to by a set name
Entity class has been updated to be its own class with ID and Name
fields. The process by which an Entity is created and retrieved from the
EntityManager now requires a Name to be supplied on creation, and then
can optionally use the Name or the ID to get the Entity. The test scene
has been updated to reflect this. The Sprite and Transform processor
classes have been updated to not refer to the Entity type (previously a
typedef for unsigned int) but instead replace all references with
"unsigned int".