Can now set texture coordinates on SpriteComponent
SpriteComponent has always had a RECT member specifying texture
coordinates, but the SpriteProcessor has not used that information. It
now does and will now scale and translate the texture coordinates
accordingly. This is done by passing a transformation matrix to the
shader. The test shader has also been updated to allow this.