Reworked Time class to use <chono> standard library
The time class now uses the portable <chrono> library as opposed to
Win32 system calls to implement the timing mechanism for the engine.
This required a small change to the existing Scene/Process class
hierarchies because the frame delta time is passed through to the scene
and also to any processes.