New implementation of SceneList
The SceneManager maintains a list of Scene types. The new implementation
stores factory methods to build instances of a scene, against the name
of a scene type. Scenes are now registered with the SceneManager in
their own implementation file via a static method, and can be referred
to by passing a string with the scene name. Also, made fixes to the
scene management logic.