Moved sound asset code to Sound subsystem
The methods used to load sound files for use in a scene has been moved
from the AssetManager class to the Sound subsystem class. Also added a
second parameter to the Load* methods, which the user can use to supply
a human readable name. That name is then passed in to the Get* method to
retrieve a pointer to the sound object.